Added experimental package 'node-yarn'
This package is mainly required as a host package to build a vue interface, but the problem is that it carries as dependencies the building of python and node, which increases the build time by many folds.

It is currently not used but kept for future reference as it was tested and it is somehow working.